htons is dead.

+++ tests/u64max.fx
@@ -2,6 +2,6 @@
 
 def main() -> int
 {
-	print((u64)0xFFFFFFFFFFFFFFFFu);
+	print(0xFFFFFFFFFFFFFFFFu);
 	return 0;
 };
